refactor(mm): Split mm_flag_mapping.rs into modular directory structure
Split the 3,823-line mm_flag_mapping.rs file into a modular directory
structure to improve maintainability and reduce git diff churn when
adding location mappings.

New structure:
- mm_flag_mapping/
  - mod.rs: Main module with static tables and re-exports
  - types.rs: MmFlagType enum
  - mapping.rs: MmFlagMapping struct
  - scenes.rs: Scene ID constants (mm_scene)
  - queries.rs: Query API functions
  - checker.rs: Status checking functions
  - tests.rs: All unit tests
  - locations/
    - mod.rs: Location registration aggregator
    - owl_statues.rs: Owl statue mappings
    - songs.rs: Song location mappings
    - masks.rs: NPC mask reward mappings
    - items.rs: Items, shops, scrubs, grottos, etc.
    - heart_pieces.rs: Heart piece locations
    - gold_skulltulas.rs: Spider house mappings
    - overworld.rs: Overworld chests and collectibles
    - dungeons/
      - mod.rs: Dungeon registration aggregator
      - main_temples.rs: Main temple chests
      - mini_dungeons.rs: Mini dungeon chests

Benefits:
- Location mappings split by category for easier navigation
- New mappings can be added without massive diffs
- Each file has clear single responsibility
- Public API unchanged - backwards compatible

Also fixes world_database doctest using incorrect region name.

Closes #640
